backend/requirements.txt
fastapi
uvicorn
groq
python-dotenv
requests
gitpython